pyproject.toml
[project]
name = "osw-selenium"
version = "0.0.1"
description = "UI testing package for OpenSemanticWorld based on Selenium and Python."
dependencies = ["selenium>=4.6.0", "python-dotenv>=1.0.0"]
authors = [
    { name = "Andreas Raeder", email = "andreas.raeder@isc.fraunhofer.de" },
]
readme = "README.md"
keywords = ['python', 'selenium', 'testing', 'opensemantic']
requires-python = ">=3.10,<4.0"
classifiers = [
    "Intended Audience :: Developers",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

[project.urls]
Homepage = "https://opensemanticworld.GitHub.io/osw-selenium/"
Repository = "https://github.com/opensemanticworld/osw-selenium"
Documentation = "https://opensemanticworld.GitHub.io/osw-selenium/"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pre-commit>=4.5.1",
    "tox-uv>=1.29.0",
    "deptry>=0.24.0",
    "ty>=0.0.14",
    "ruff>=0.14.14",
    "sphinx>=8.1.3",
    "sphinx-autobuild>=2024.10.3",
    "furo",
    "sphinx-copybutton",
    "sphinx-autodoc-typehints",
    "myst-parser",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/osw_selenium"]

[tool.ty.environment]
python = "./.venv"
python-version = "3.10"

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "integration: tests that require a running OSW instance (MW_SITE_SERVER)",
]

[tool.ruff]
target-version = "py310"
line-length = 120
fix = true

[tool.ruff.lint]
select = [
    # flake8-2020
    "YTT",
    # flake8-bandit
    "S",
    # flake8-bugbear
    "B",
    # flake8-builtins
    "A",
    # flake8-comprehensions
    "C4",
    # flake8-debugger
    "T10",
    # flake8-simplify
    "SIM",
    # isort
    "I",
    # mccabe
    "C90",
    # pycodestyle
    "E",
    "W",
    # pyflakes
    "F",
    # pygrep-hooks
    "PGH",
    # pyupgrade
    "UP",
    # ruff
    "RUF",
    # tryceratops
    "TRY",
]
ignore = [
    # LineTooLong
    "E501",
    # DoNotAssignLambda
    "E731",
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["S101", "S105", "S106", "TRY003"]

[tool.ruff.format]
preview = true
